add beginex and change float64 to int64

+++ internal/db/init/001_init.sql
@@ -1,13 +1,13 @@
 CREATE TABLE IF NOT EXISTS accounts (
 	account_id BIGINT PRIMARY KEY,
-	balance NUMERIC(10, 4) NOT NULL DEFAULT 0
+	balance BIGINT NOT NULL DEFAULT 0
 );
 
 CREATE TABLE IF NOT EXISTS transactions (
 	id BIGSERIAL PRIMARY KEY,
 	source_account_id BIGINT REFERENCES accounts(account_id),
   destination_account_id BIGINT REFERENCES accounts(account_id),
-	amount NUMERIC(10, 4) NOT NULL DEFAULT 0,
+	amount BIGINT NOT NULL DEFAULT 0,
 	created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP
 );
